dashboard/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
cffi==1.15.1
    # via cryptography
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via flask
cryptography==40.0.2
    # via -r requirements.in
dash==2.9.2
    # via
    #   -r requirements.in
    #   dash-auth
dash-auth==2.0.0
    # via -r requirements.in
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
distlib==0.3.6
    # via virtualenv
dnspython==2.3.0
    # via pymongo
filelock==3.11.0
    # via virtualenv
flask==2.2.3
    # via
    #   -r requirements.in
    #   dash
gevent==22.10.2
    # via -r requirements.in
greenlet==2.0.2
    # via gevent
gunicorn==20.1.0
    # via -r requirements.in
identify==2.5.22
    # via pre-commit
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
mysql-connector==2.2.9
    # via -r requirements.in
nodeenv==1.7.0
    # via pre-commit
numpy==1.24.2
    # via
    #   -r requirements.in
    #   pandas
packaging==23.0
    # via plotly
pandas==2.0.0
    # via -r requirements.in
platformdirs==3.2.0
    # via virtualenv
plotly==5.14.1
    # via
    #   -r requirements.in
    #   dash
pre-commit==3.2.2
    # via -r requirements.in
pycparser==2.21
    # via cffi
pymongo==4.3.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via pre-commit
six==1.16.0
    # via python-dateutil
tenacity==8.2.2
    # via plotly
tzdata==2023.3
    # via pandas
virtualenv==20.21.0
    # via pre-commit
werkzeug==2.2.3
    # via flask
zope-event==4.6
    # via gevent
zope-interface==6.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
